# apps/mas-ops-api/openapi.json
{
  "components": {
    "schemas": {
      "ActivityEventResponse": {
        "description": "Incident activity timeline entry.",
        "properties": {
          "activity_id": {
            "title": "Activity Id",
            "type": "integer"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "source_event_id": {
            "title": "Source Event Id",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "activity_id",
          "source_event_id",
          "client_id",
          "fabric_id",
          "incident_id",
          "asset_id",
          "event_type",
          "subject_type",
          "subject_id",
          "payload",
          "occurred_at"
        ],
        "title": "ActivityEventResponse",
        "type": "object"
      },
...
      "AssetResponse": {
        "description": "Asset detail response.",
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "asset_kind": {
            "title": "Asset Kind",
            "type": "string"
          },
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "health_observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Observed At"
          },
          "health_state": {
            "title": "Health State",
            "type": "string"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "last_alert_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Alert At"
          },
          "mgmt_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mgmt Address"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "vendor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor"
          }
        },
        "required": [
          "asset_id",
          "client_id",
          "fabric_id",
          "asset_kind",
          "vendor",
          "model",
          "hostname",
          "mgmt_address",
          "site",
          "tags",
          "health_state",
          "health_observed_at",
          "last_alert_at",
          "updated_at"
        ],
        "title": "AssetResponse",
        "type": "object"
      },
      "ChatMessageCreateRequest": {
        "description": "Append-message request body.",
        "properties": {
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatMessageCreateRequest",
        "type": "object"
      },
...
      "ChatScope": {
        "description": "Supported operator chat scopes.",
        "enum": [
          "global",
          "incident"
        ],
        "title": "ChatScope",
        "type": "string"
      },
      "ChatSessionCreateRequest": {
        "description": "Create-chat-session request body.",
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "fabric_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fabric Id"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ChatScope"
          }
        },
        "required": [
          "scope"
        ],
        "title": "ChatSessionCreateRequest",
        "type": "object"
      },
      "ChatSessionResponse": {
        "description": "Chat session with turns and messages.",
        "properties": {
          "chat_session_id": {
            "title": "Chat Session Id",
            "type": "string"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_user_id": {
            "title": "Created By User Id",
            "type": "string"
          },
          "fabric_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fabric Id"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "scope": {
            "$ref": "#/components/schemas/ChatScope"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/ChatTurnResponse"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "chat_session_id",
          "scope",
          "client_id",
          "fabric_id",
          "incident_id",
          "created_by_user_id",
          "created_at",
          "turns",
          "messages"
        ],
        "title": "ChatSessionResponse",
        "type": "object"
      },
      "ChatTurnResponse": {
        "description": "Persisted chat turn response.",
        "properties": {
          "actor_user_id": {
            "title": "Actor User Id",
            "type": "string"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "chat_session_id": {
            "title": "Chat Session Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ChatTurnState"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "turn_id": {
            "title": "Turn Id",
            "type": "string"
          }
        },
        "required": [
          "turn_id",
          "request_id",
          "chat_session_id",
          "actor_user_id",
          "state",
          "submitted_at",
          "completed_at",
          "approval_id"
        ],
        "title": "ChatTurnResponse",
        "type": "object"
      },
...
      "ClientSummaryResponse": {
        "description": "Client summary read-model response.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "critical_asset_count": {
            "title": "Critical Asset Count",
            "type": "integer"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "open_alert_count": {
            "title": "Open Alert Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "fabric_id",
          "name",
          "open_alert_count",
          "critical_asset_count",
          "updated_at"
        ],
        "title": "ClientSummaryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentResponse": {
        "description": "Incident detail response.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "client_id",
          "fabric_id",
          "state",
          "severity",
          "summary",
          "opened_at",
          "updated_at"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MAS Ops API",
    "version": "0.4.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/assets/{asset_id}": {
      "get": {
        "description": "Fetch one asset detail.",
        "operationId": "get_asset_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset",
        "tags": [
          "assets"
        ]
      }
    },
    "/assets/{asset_id}/activity": {
      "get": {
        "description": "Fetch one asset's projected visibility activity.",
        "operationId": "list_asset_activity_assets__asset_id__activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityEventResponse"
                  },
                  "title": "Response List Asset Activity Assets  Asset Id  Activity Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Asset Activity",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/chat/sessions": {
      "post": {
        "description": "Create a global or incident-scoped chat session.",
        "operationId": "create_chat_session_chat_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Chat Session",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/sessions/{chat_session_id}": {
      "get": {
        "description": "Fetch one chat session and its persisted turns/messages.",
        "operationId": "get_chat_session_chat_sessions__chat_session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_session_id",
            "required": true,
            "schema": {
              "title": "Chat Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chat Session",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/sessions/{chat_session_id}/messages": {
      "post": {
        "description": "Append one operator message to a chat session.",
        "operationId": "append_chat_message_chat_sessions__chat_session_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_session_id",
            "required": true,
            "schema": {
              "title": "Chat Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Chat Message",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/clients/{client_id}": {
      "get": {
        "description": "Fetch one authorized client summary.",
        "operationId": "get_client_clients__client_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Client",
        "tags": [
          "clients"
        ]
      }
    },
    "/clients/{client_id}/activity": {
      "get": {
        "description": "List the latest visibility activity for one authorized client.",
        "operationId": "list_client_activity_clients__client_id__activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityEventResponse"
                  },
                  "title": "Response List Client Activity Clients  Client Id  Activity Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Client Activity",
        "tags": [
          "clients"
        ]
      }
    },
    "/clients/{client_id}/assets": {
      "get": {
        "description": "List assets for one authorized client.",
        "operationId": "list_client_assets_clients__client_id__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  },
                  "title": "Response List Client Assets Clients  Client Id  Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Client Assets",
        "tags": [
          "clients"
        ]
      }
    },
...
    "/clients/{client_id}/incidents": {
      "get": {
        "description": "List incidents for one authorized client.",
        "operationId": "list_client_incidents_clients__client_id__incidents_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response List Client Incidents Clients  Client Id  Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Client Incidents",
        "tags": [
          "clients"
        ]
      }
    },
...
  }
}